from cs

+++ resources/views/staticPages/users.blade.php
@@ -10,7 +10,7 @@
       <a href="{{ route('user.profile', $user->id) }}">
         <div class="card-body">
           {{ $user->name }}さんのページへ。
-          @if (Auth::user()->admin == 1)
+          @if (Auth::user()->admin == 1 && Auth::user()->id != $user->id)
             {!! Form::open(['action' => ['UsersController@destroy', $user->id], 'method' => 'DELETE', 'id' => 'form_'.$user->id]) !!}
             {{ method_field('delete') }}
               <input type="submit" value="削除" data-id="{{ $user->id }}"  style="display:inline-block;background:transparent;border:none;" onclick="deleteSubmit(this)">
